+++ application/views/email/verify_email.php
@@ -1,0 +1,3 @@
+<p>Bienvenido a Archmage.</p>
+<p>Verifica tu correo pulsando este enlace:</p>
+<p><a href="<?php echo htmlentities($link); ?>"><?php echo htmlentities($link); ?></a></p>